added tests for flexible_partial()
