+++ dask_grblas/base.py
@@ -1,15 +1,14 @@
 from numbers import Number
 from functools import partial
-from reprlib import recursive_repr
 import dask.array as da
 import grblas as gb
 import numpy as np
-from scipy.sparse import csr_matrix
 from grblas.operator import UNKNOWN_OPCLASS, find_opclass, get_typed_op
 from grblas.dtypes import lookup_dtype
 
 from . import replace as replace_singleton
 from .mask import Mask
+from .functools import flexible_partial
 from .utils import pack_args, pack_kwargs, get_grblas_type, get_meta, np_dtype, wrap_inner
 
 _expect_type = gb.base._expect_type
@@ -426,8 +425,11 @@ class DOnion:
         :packed_args: a list of positional arguments to `func`
         :packed_kwargs: a dict of named arguments to `func`
         """
-        omit_DOnion = is_DOnion
-        func = flexible_partial(func, omit_DOnion, *packed_args, **packed_kwargs)
+        where_non_DOnion = [False if is_DOnion(arg) else True for arg in packed_args]
+        non_Donions = [arg for arg in packed_args if not is_DOnion(arg)]
+        non_DOnion_kwargs = {k: v for (k, v) in packed_kwargs.items() if not is_DOnion(v)}
+        func = flexible_partial(func, where_non_DOnion, *non_Donions, **non_DOnion_kwargs)
+
         donion_args = tuple(arg.kernel for arg in packed_args if is_DOnion(arg))
         donion_kwargs = {k: v.kernel for (k, v) in packed_kwargs.items() if is_DOnion(v)}
         kernel = da.map_blocks(func, *donion_args, **donion_kwargs, dtype=dtype, meta=meta)
@@ -453,182 +455,6 @@ class DOnion:
 
 
 is_DOnion = partial(is_type, DOnion)
-
-
-class skip:
-    def __repr__(self):
-        return "skip"
-    __str__ = __repr__
-    __reduce__ = __repr__  # This makes it pickle well!
-
-skip = skip()
-
-
-def normalize_occupancies(specs):
-    # Converts any valid `specs` to the form: [True, False, True, ...]
-    if isinstance(specs, Iterable):
-        try:
-            a = np.asarray(specs)
-        except Exception as e:
-            raise e
-
-        if a.ndim == 1:
-            if a.dtype is np.bool_:
-                return a, None
-            
-            if a.dtype.kind in np.typecodes["AllInteger"]:
-                pos = csr_matrix(np.ones_like(a, dtype=np.bool_), a, np.array([0]))
-                return pos.toarray(), None
-
-            occupancy = np.array([False if x is skip else True for x in a])
-            args = [x for x in a if x is not skip]
-            return occupancy, args
-
-        if a.ndim == 2 and a.shape[1] == 2:
-            pos = a[:, 0]
-            pos = csr_matrix(np.ones_like(pos, dtype=np.bool_), pos, np.array([0]))
-            return pos.toarray(), a[:, 1]
-
-    raise ValueError(
-        'specs should be an iterable of any of the following forms:\n'
-        '[True, False, True, ...]\n'
-        '[0, 2, 3, ...]\n'
-        '(skip, b, skip, skip, d, ...)'
-        '((1, b), (3, d), ...)'
-    )
-
-
-################################################################################
-### flexible_partial() argument application
-################################################################################
-
-# Purely functional, no descriptor behaviour
-class flexible_partial:
-    """New function with flexible partial application of the given
-    arguments and keywords.
-    """
-
-    __slots__ = "func", "args", "vacancies", "kwargs", "__dict__", "__weakref__"
-
-    def __new__(cls, func, specs, /, *args, **kwargs):
-        # Validate input parameters:
-        if not callable(func):
-            raise TypeError("the first argument must be callable")
-
-        occupancies, args_ = normalize_occupancies(specs)
-        args = args if args_ is None else args_
-        new_arg = iter(args)
-
-        nfilled = np.count_nonzero(occupancies)
-        nargs = len(args)
-        if nargs != nfilled:
-            raise ValueError(
-                f"Number ({nargs}) of given arguments does not match "
-                f"number ({nfilled}) of argument slots to be occupied."
-            )
-
-        self = super(flexible_partial, cls).__new__(cls)
-        
-        if hasattr(func, "func") and hasattr(func, "vacancies") and hasattr(func, "kwargs"):
-            func = func.func
-            self.args = list(func.args)
-            kwargs = {**func.kwargs, **kwargs}
-
-            old_vacancy = iter(func.vacancies)
-            slot_status = iter(occupancies)
-
-            # step through old vacancies:
-            for occupy in occupancies:
-                try:
-                    pos = next(old_vacancy)
-                except StopIteration:
-                    # old vacancies now exhausted => reset occupancies and continue elsewhere:
-                    occupancies = list(slot_status)
-                    break
-                else:
-                    next(slot_status)
-
-                if occupy:
-                    # fill the vacancy:
-                    self.args[pos] = next(new_arg)
-                else:
-                    # record the vacancy:
-                    self.vacancies.append(pos)
-        else:
-            self.args = []
-            self.vacancies = []
-
-        start = len(self.args)
-        for pos, occupied in enumerate(occupancies, start=start):
-            if occupied:
-                self.args.append(next(new_arg))
-            else:
-                # create a vacancy:
-                self.args.append(None)
-                self.vacancies.append(pos)
-
-        self.func = func
-        self.args = tuple(self.args)
-        self.vacancies = tuple(self.vacancies)
-        self.kwargs = kwargs
-        return self
-
-    def __call__(self, /, *args, **kwargs):
-        if len(args) < len(self.vacancies):
-            raise ValueError(f"Expected at least {len(self.vacancies)} positional arguments. "
-                             f"Got {len(args)}.")
-
-        new_arg = iter(args)
-        self_args = list(self.args)
-        for pos in self.vacancies:
-            # fill all the vacancies
-            self_args[pos] = next(new_arg)
-
-        # append the remaining arguments and make the call:
-        self_args.extend(new_arg)
-        kwargs = {**self.kwargs, **kwargs}
-        return self.func(*self_args, **kwargs)
-
-    @recursive_repr()
-    def __repr__(self):
-        qualname = type(self).__qualname__
-        args = [repr(self.func)]
-        c = count()
-        args.extend("_" if next(c) in self.vacancies else repr(x) for x in self.args)
-        args.extend(f"{k}={v!r}" for (k, v) in self.kwargs.items())
-        if type(self).__module__ == "functools":
-            return f"functools.{qualname}({', '.join(args)})"
-        return f"{qualname}({', '.join(args)})"
-
-    def __reduce__(self):
-        return type(self), (self.func,), (self.func, self.args, self.vacancies,
-               self.kwargs or None, self.__dict__ or None)
-
-    def __setstate__(self, state):
-        if not isinstance(state, tuple):
-            raise TypeError("argument to __setstate__ must be a tuple")
-        if len(state) != 5:
-            raise TypeError(f"expected 5 items in state, got {len(state)}")
-        func, args, vacs, kwds, namespace = state
-        if (not callable(func) or not isinstance(args, tuple) or not isinstance(vacs, tuple) or
-           (kwds is not None and not isinstance(kwds, dict)) or
-           (namespace is not None and not isinstance(namespace, dict))):
-            raise TypeError("invalid flexible_partial state")
-
-        args = tuple(args) # just in case it's a subclass
-        vacs = tuple(vacs)
-        if kwds is None:
-            kwds = {}
-        elif type(kwds) is not dict: # XXX does it need to be *exactly* dict?
-            kwds = dict(kwds)
-        if namespace is None:
-            namespace = {}
-
-        self.__dict__ = namespace
-        self.func = func
-        self.args = args
-        self.vacancies = vacs
-        self.keywords = kwds
 
 
 # Dask task functions
